Remove unnecessary QuestionPresenter#has_suffix_label? method
Previously `MoneyQuestionPresenter#has_suffix_label?` meant that a money
question never rendered a suffix label. By overriding
`QuestionPresenter#suffix_label` and returning `nil`, we can make the behaviour
more explicit while simplifying the code.

In the case where the content inside the condition is *only* the `suffix_label`
itself, it's safe to remove the condition and rely on the fact that a `nil` or empty `String` will have no effect on the rendered HTML.

In the case where there is more than just the `suffix_label` inside the
condition, it's more idiomatic Rails to use `Object#present?` on the
`suffix_label`.