Generate a schema-valid content-item
The `FlowContentItem` class generates a payload that we will send to
the content-store (via the publishing-api).

We use the `govuk-content-schemas` to verify what what we're building
here is a valid object.

To run these schema tests, there needs to be a copy of `govuk-content-schemas`
in a sibling directory to this project (`../govuk-content-schemas`). To provide
this during tests on CI we check out the repo in jenkins.sh. The code for this
is copied from the collections-publisher repo because it works there.

You can read more about setting up contract tests here:

https://github.com/alphagov/govuk-content-schemas/blob/master/docs/contract-testing-howto.md