Move landing page content to ERB template
Since creating this branch we've switching to storing landing page content in
ERB templates. I've recently rebased this branch on master which means that I
now need to store the landing page content for this Smart Answer in ERB rather
than YAML. Ideally I would've amended an earlier commit so that we never added
the content to YAML in the first place. Unfortunately I think that's going to be
quite a lot of work and not necessarily worth the effort.