Avoid displaying 0000 in previous question responses
We use a `date_question` in calculate-your-part-year-profit to collect a day and
month from the user. We do this by setting the `default_year` to 0. This works
as expected but has two slightly undesirable side effects:

* The date in the URL includes the four digit year as "0000".

* The date in the list of Previous answers includes the four digit year as
  "0000".

The date in the URL probably isn't quite so important as I suspect it won't even
be seen/noticed by the majority of users.

The date in the list of Previous answers is slightly more obvious and has been
picked up on in HMRC's feedback of the calculate-your-part-year-profit Smart
Answer. This incredibly naive implementation simply strips " 0000" from the date
to be presented to the user.

I considered introducing a `day_and_month_question` but I'm not convinced it's
necessarily worth the effort at this time.