Add rate for ages over max age in pension_top_up_data.yml
Previously we were incorrectly informed that state pension top up is available
to users up to 100 years old. Actually the rule is that once users reach 100
and above, their top up rate stops decreasing.

So we want the remove the "too old" limit for state pension top up. This is a
step in that direction, but note that we haven't yet made changed the externally
visible behaviour.