Remove "too old" limit for state pension top up
Previously we were incorrectly informed that state pension top up is available
to users up to 100 years old. Actually the rule is that once users reach 100
and above, their top up rate stops decreasing.

This commit removes the check for the user being too old to be able to top-up
their pension and the subsequent outcome along with its content and tests.