Fix regression test artefacts for state-pension-topup
This should've been part of a previous commit [1].

I forgot to run the regression tests after I'd rebased the branch against
master. Rebasing against master brought in the changes which added regression
test artefacts for question pages and of course these were generated *before*
the changes in this branch. Thus the question page artefacts need to be
updated which is what this commit achieves.

[1]: b01001a8c218227ebe903f64ccc663c84c8db7cb