Delete animal-example-simple example smartdown flow
The plan is to phase out Smartdown flows. This flow is marked as "draft" so it
doesn't currently appear in production. Also it's clearly not intended to ever
be deployed to production. So I think it's safe to remove it.

I've also removed the flow's plugins.

Note that some tests refer to a Smartdown flow with the same name, but that flow
is defined in `test/fixtures/smartdown_flows`.