Delete animal-example-multiple example smartdown flow
The plan is to phase out Smartdown flows. This flow is marked as "draft" so it
doesn't currently appear in production. Also it's clearly not intended to ever
be deployed to production. So I think it's safe to remove it.

I've also removed the flow's plugin and its corresponding unit test.