Make employee AWE variable available to next_node blocks
I originally made this change so that we could move the `:usual_work_days?`
question earlier in the flow. However, it turned out that in the end we didn't
need to move the question.

I've decided that, although it's not strictly necessary, I'm still going to
make this change so I can more easily apply some refactoring commits. I don't
think the change does any harm.