Make employee AWE variable available to next_node blocks
We want to move the `:usual_work_days?` question earlier in the flow. This means
we'll need to instantiate the `OutcomeDecision` class in the `next_node` blocks
of these three questions. That class depends on the following state variable:

    employee_average_weekly_earnings

By converting the `calculate` block to a `next_node_calculation` block we can
make the variable available.