Implement QuestionPresenter#post_body
This is in preparation for converting Smartdown Smart Answers to Ruby.

Smartdown includes the ability to display a `post_body` (rendered as Govspeak)
after the form on a question page. This is displayed by the
"_current_question.html.erb" partial.

This commit adds support for `post_body` to Ruby Smart Answers.