Remove unused conditionals from mat-shared-pay.txt
The mat-shared-pay snippet is rendered by the following outcomes:

* outcome_mat-leave_mat-pay_mat-shared-leave_mat-shared-pay
* outcome_mat-leave_mat-pay_pat-leave_both-shared-leave_mat-shared-pay
* outcome_mat-pay_mat-shared-pay
* outcome_mat-pay_pat-leave_pat-shared-leave_mat-shared-pay

These outcomes are all reached from employment_details_of_partner_2b, which can
only be reached from employment_details_of_partner_2a, which in turn can only be
reached if the due date is on or after 5 Apr 2015. This means that the check for
`$IF range_in_2013_2014_fin_year?(due_date)` will never be true.