Fix question number comments
I'm not convinced these question numbers are very useful. I assume they're
related to question numbers in the logic document, but don't know whether
the logic document has been kept up-to-date and it's obvious from the changes
in this commit that the comments haven't been kept up-to-date. My inclination
is to delete them, but for now I've just fixed them as best I can so they are
at least consistent in the context of the flow class.