Only run pay-leave-for-parents tests when required
These tests are in place ready for the conversion of this Smart Answer from
Smartdown to Ruby. We don't need them to run as part of the main regression test
suite until the Smart Answer has been converted.

This isn't the best implementation but I think it's probably OK given that I'll
revert it as soon as this Smart Answer has been converted to Ruby.