Update YAML files using current YAML version
The version of YAML has changed since these files were generated and the newer
version seems to generate slightly different YAML output e.g. it uses single-
quotes instead of double-quotes.

Because of this, when we make a change and regenerate these files, it's
difficult to see the woods for the trees. By loading and saving all the YAML
files in this commit, I hope that it will be easier to see what's changed in
later commits.

Some of the `*-questions-and-responses.yml` files included comments which
would've been lost in this automated process. So I ran the following command
to identify them:

    grep -R "#" test/data/*

I then ran the following command to do the automatic update:

    rails r script/update-yaml.rb

And finally I manually edited the files to re-add the comments identified
above. I re-ran the `grep` command to check I hadn't missed any comments.