Update regression test artefacts after moving usual_work_days Q
Theoretically it feels as if this should just be a bunch of renames of outcome
page artefacts where the path changes to reflect the different order of
questions. However, it seems as if `git` diff doesn't manage to work out the
renames correctly.

I also expect to _both_ the path and the content of question page artefacts
change. The path should've changed for the same reason as the outcome pages
and the content will have changed because of things like the 'Previous answers'
section changing.

Since it's proved very difficult to get confidence that these changes as are I
expect from looking at the diff, I've used the `visualise` action to generate
a diagram of the flow with the connections between its questions and outcomes
before and after the moving of the `:usual_work_days?` question. This has
convinced me that these changes are almost certainly correct.