Only consider number of days *worked* in linked sickness period
This fixes a bug reported in Zendesk ticket #1128193 [1].

I've added a new integration test based on one of the examples which have been
fact-checked by HMRC.

I've updated the assertion in three other existing integration tests to match
the new table output, but I don't have complete confidence that these changes
are correct.

The changes to the first row in the table in each case makes sense to me,
because presumably there's been a reduction in the number of working days
in the linked sickness period, meaning that it takes longer to use up the
three "waiting days", and hence the SSP amount for the first week is reduced.

However, the removal of a number of rows towards the end of the table makes
less sence. Presumably this is something to do with hitting the 28-week (?)
maximum, but it's not obvious to me why we would hit the maximum sooner. I
plan to check this with @lutgendorff before merging.

I'm not completely happy with the code change, but the rest of the logic is so
hard to follow that I've purposefully tried to keep the scope of the change as
small as possible.

Really this smart answer needs some major refactoring, but that's going to have
to wait for now.

[1]: https://govuk.zendesk.com/tickets/1128193