Only consider number of days *worked* in linked sickness period
This fixes a bug reported in Zendesk ticket #1128193 [1].

I've added a new integration test based on one of the examples which have been
fact-checked by HMRC.

I've had to update a couple of other integration tests to take into account the
new behaviour:

For the "no SSP payable as already had maximum" context, I had to increase the
length of the linked sickness period to ensure that the user still reached the
maximum entitlement given that now only Mon-Fri days in the linked period of
sickness count as qualifying days.

For the other one, I had to reduce the SSP amount for the first week, because
there's been a reduction in the number of _working_ days in the linked sickness
period, meaning that it takes longer to use up the three "waiting days", and
hence the SSP amount for the first week is reduced.

I've also updated one of the calculator unit tests, because the initial state
of the calculator no longer made any sense - since there are *no* days in the
working pattern, it's now impossible for there to be any `prev_sick_days`.

I'm not completely happy with the code change, but the rest of the logic is so
hard to follow that I've purposefully tried to keep the scope of the change as
small as possible.

Really this smart answer needs some major refactoring, but that's going to have
to wait for now.

[1]: https://govuk.zendesk.com/tickets/1128193