Only consider number of days *worked* in linked sickness period
This fixes a bug reported in Zendesk ticket #1128193 [1].

I've added a new integration test based on one of the examples which have been
fact-checked by HMRC.

I've had to update a couple of other tests to take into account the new
behaviour:

For the "no SSP payable as already had maximum" context, I had to increase the
length of the linked sickness period to ensure that the user still reached the
maximum entitlement.

For the other one, I had to reduce the SSP amount for the first week, because
there's been a reduction in the number of _working_ days in the linked sickness
period, meaning that it takes longer to use up the three "waiting days", and
hence the SSP amount for the first week is reduced.

I'm not completely happy with the code change, but the rest of the logic is so
hard to follow that I've purposefully tried to keep the scope of the change as
small as possible.

Really this smart answer needs some major refactoring, but that's going to have
to wait for now.

[1]: https://govuk.zendesk.com/tickets/1128193