Add qualifying week start to maternity/paternity calculator.
This change updates the maternity, paternity and adoption smart answer calculators by asking the user if they were, or will be, employed during the qualifying week - rather than the present date when they are completing the questions.

The qualifying week (QW) is, broadly, the 15th week before the expected week of childbirth or adoption. It's worked out by finding the Sunday before the due or match date (or due day itself if it's a Sunday). The QW is 15 Sundays back from there.

For example, if the due date is 1 January 2016, the QW is 13 September.