Fail fast if an ArgumentError is raised in a next_node block
I don't think it's a good idea to rescue from such a generic exception in our
`next_node` blocks. It hides real problems in our Ruby code and makes them look
as though the user has entered invalid data in the Smart Answer.

The rescuing of ArgumentError caused me problems twice recently. Once when I was
comparing a Date with a String and once when I was calling a method with
incorrect arguments. This change will mean that errors such as this cause the
application to fail fast.

    >> Date.today < '2016-01-01'
    ArgumentError: comparison of Date with String failed

    >> def foo(arg); end
    => :foo
    >> foo()
    ArgumentError: wrong number of arguments (0 for 1)

The two preceeding commits ("Raise InvalidResponse when parsing Integers in
Value#parse_input" and "Raise InvalidResponse when parsing Floats in
Value#parse_input") should have removed the only places where we were raising
an `ArgumentError` in response to bad user input.