Raise InvalidResponse when parsing Floats in Value#parse_input
This is in preparation for removing the `rescue ArgumentError` statement from
`SmartAnswer::Flow#process`.

Rescuing `ArgumentError` makes it easy to mask real exceptions in the code in
our `next_node` blocks.

I think one of the legitimate reasons to rescue `ArgumentError` at the moment is
because they can be raised by bad user input in this method. Raising
`InvalidResponse` instead seems like the right thing to do and should allow me
to stop rescuing `ArgumentError`.