Add validation for SSP linked sickness end Q
The need for this was identified during the fact-checking process.

The linked sickness period must be *separate* from the current sickness period
otherwise it would count as part of the current sickness period.

I've also updated the regression test checksums - validation errors aren't
typically tested in the regression tests so, as expected, there are no changes
to the test artefacts.