Add validation for SSP linked_sickness_end_date? question
The need for this was identified during the fact-checking process.

The linked sickness period must be *separate* from the current sickness period
otherwise it would count as part of the current sickness period. This means
there must be at least a 1-day gap between the end of the linked period of
sickness and the start of the current period of sickness.