Make SmartdownAdapter::Presenter#page_title a noop
This removes the `h2.page-title` element from appearing on a question page.
Specifically, it means that the conditional (`<% if @presenter.page_title %>`)
on line 20 of the _content.html.erb partial is never true.

This mirrors the behaviour in Ruby Smart Answers and will (hopefully) make it
easier to convert pay-leave-for-parents from Smartdown to Ruby. This is because
it'll be easier to see any differences in the regression test output if we're
expecting them to be identical.

I considered implementing SmartAnswerPresenter#page_title instead but don't
think the concept of a page title is so useful now that we only have single
questions per page.

I've regenerated the test artefacts for pay-leave-for-parents (the only
remaining Smartdown flow) as part of this commit as there weren't many changes.