Make PreviousQuestionPagePresenter#title a noop
The effect of this change is to remove the question page title from the
"Previous answers" section that appears at the bottom of each question page.
Specifically, it means that the first conditional (`<%- if question_page.title -%>`)
in the _collapsed_question.html.erb partial is never true.

This mirrors the behaviour in Ruby Smart Answers and will (hopefully) make it
easier to convert pay-leave-for-parents from Smartdown to Ruby. This is because
it'll be easier to see any differences in the regression test output if we're
expecting them to be identical.

I considered implementing this same functionality in Ruby Smart Answers instead
but don't think the concept of a page title is so useful now that we only have
single questions per page.

I've regenerated the test artefacts for pay-leave-for-parents (the only
remaining Smartdown flow) as part of this commit as there weren't many changes.