Change regression test responses for SSP
In preparation for adding a new validation to the `:linked_sickness_end_date?`
question, I've changed a couple of the responses for the regression test.

The validation will ensure that this date is before the answer to the
`:first_sick_day?` question. Unfortunately, some combinations of the regression
test responses will not pass this validation.

This commit changes a couple of the responses so that the validation will pass
and updates the regression test artefacts accordingly. The changes to the
artefacts should all effectively be renames, but unfortunately it looks as if
`git`'s heuristics haven't correctly identified all of the renames.

I've also updated the regression test checksums to reflect the changes to the
responses.