Change regression test responses for SSP
This is in preparation for adding a new validation rule to the
`:linked_sickness_end_date?` question. The validation rule will ensure that
this date is before the answer to the `:first_sick_day?` question.

The combinations of the existing regression test responses including the
`linked_sickness_end_date?` response of `2013-09-01` would not pass this
validation rule.

This commit changes this response from `2013-09-01` to `2013-03-30` so
that all combinations will now pass the new validation rule.

Based on the new questions-and-responses file, I've regenerated the
responses-and-expected-results file using the following command:

    $ rails r script/generate-responses-and-expected-results-for-smart-answer.rb calculate-statutory-sick-pay