Remove 'Ordinary' from 'Ordinary Statutory Paternity Pay' in SSP
This change came to light during the fact-checking process.

Apparently we should now refer to it as just:

    Statutory Paternity Pay

I've also updated the relevant question option key, even though this will make
life more painful with the regression tests. I did this because the options
key appears in the URL path and I thought it would look odd if it still
included the word, `ordinary`.

I plan to update the regression tests in the next commit to make it easier to
see what's changed.