Change integration test responses for SSP
In preparation for adding a new validation to the `:linked_sickness_end_date?`
question, I've changed a few of the responses for the integration test.

The validation will ensure that this date is before the answer to the
`:first_sick_day?` question. Unfortunately, some of the integration test
responses will not pass this validation.

This commit changes the relevant responses so that the validation will pass
while leaving the outcome result unchanged. This should mean that these tests
will continue to pass even after the new validation is added.