Rubocop fixes
I decided against amending the commits which triggered these Rubocop violations,
because (a) in many cases it would've made it harder to see a refactoring; and
(b) because it would've meant amending a bunch of commits where the regression
test checksums are re-generated.

I've updated the regression test checksums - there were no changes to the
regression test artefacts.