Move pay-leave-for-parents templates to outcome directory
Outcomes have been moved to their own directory since this branch was first
created.

I considered updating the Smartdown conversion script to save the outcomes in
the correct directory, and then rewriting history to avoid having this commit. I
didn't do that because I've made a number of changes to the outcomes after the
conversion and I think it would be painful to apply those changes if the
outcomes exist in a new directory.