Convert StatutorySickPayCalculator into ActiveModel model
This will allow us to instantiate the calculator at the beginning of the flow
and set its attributes as we go along c.f. `PartYearProfitTaxCreditsCalculator`.

This commit extracts logic from the constructor into separate methods which are
only called on demand. This should make it easier to unit test these methods,
because we'll only need to supply the attributes that are actually needed for
the method under test; previously we had to supply many/all of the attributes to
avoid an error in the constructor.