Use next_node with a block in SSP Q5
I want to be able to assign the calculator in the first question and then
assign attributes as they become available. In the most recently added
Smart Answer, `PartYearProfitTaxCreditsFlow`, we used `next_node` blocks to
do this assignment. Also more generally we've decided to deprecate the
`next_node_if` / predicate style.