Avoid unnecessary negative in boolean method on SSP calculator
Inverting the logic of the `#not_getting_maternity_pay?` (renaming it to
`#already_getting_maternity_pay?`) and reversing the if/else statements that
use it makes the code more readable.

The name of the method still seems a bit at odds with the implementation, but
I still think it's an improvement.