Move prior_sick_days calculation into SSP calculator
We no longer need to pass in the value for the `prev_sick_days` attribute,
because we can calculate it from attributes already available on the calculator.

The `#prev_sick_days` & `#prior_sick_days` methods represent the same thing, but
to keep the changes in this commit to a minimum, I've kept both methods with one
calling the other.

The defaulting of `#prev_sick_days` is now taken care of with a guard clause in
`#prior_sick_days`.

Unfortunately I've had to update most of the calculator unit tests to set the
3 attributes relating to the linked sickness period instead of setting
`prev_sick_days`. In each case I've added an assertion to check that the
attribute values I've chosen result in the same value of `#prev_sick_days`.
These assertions are just to check the setup and they could probably be removed,
but I found them helpful in getting confidence in this change.

In general there are way too many of these tests - I'm confident that the
same behaviour could be tested much more simply - but that's going to have to
wait for another day!