Convert SSP Q8 response into calculator attributes
This is another small step in pushing as much policy logic into the calculator
as possible.

Also moved `#pay_day_offset` calculation block which depended on the response
into method on calculator.

I had to add a `precalculate` block for `:relevant_period_to` to the
`:total_employee_earnings?` question so that the value from the calculator
can be interpolated in the question title in the i18n YAML file.