Convert SSP Q8.1 response into calculator attributes
This is another small step in pushing as much policy logic into the calculator
as possible.

I've consciously removed the comment relating to the `relevant_period_from`
block, because I don't think it adds anything over and above the code itself.

I've kept the mysterious `+ 1.day` so that the behaviour should be unchanged.
I note that this addition also appears in the logic document without
explanation, so I've asked for clarification.

Also moved `monthly_pattern_payments` calculation block which depended on the
response into method on calculator.

I had to add a `precalculate` block for `:relevant_period_from` to the
`:total_employee_earnings?` question so that the value from the calculator
can be interpolated in the question title in the i18n YAML file.