Extract AWE calculation into calculator
The following responses are now stored on the calculator and the
`employee_average_weekly_earnings` attribute has become a method in its own
right.

* `total_employee_earnings`
* `contractual_days_covered_by_earnings`
* `days_covered_by_earnings`

Note that the `#employee_average_weekly_earnings` method chooses the relevant
calculation based on the response to the `:paid_at_least_8_weeks?` question.

Note also that I've had to introduce a `precalculate` block in the
`:not_earned_enough` outcome in order to make the value available to the
outcome template.