Move formatting of SSP weekly payments into outcome template
* I removed the `#formatted_sick_pay_weekly_amounts` method from the
`SmartAnswer::Calculators::StatutorySickPayCalculator` and made the
`#weekly_payments` method public to take its place for use in the outcome
template.

* This in turn meant that I needed to change some of the calculator unit tests
to call the `#weekly_payments` method instead and to change the assertions
accordingly.

* I've used `SmartAnswer::OutcomeHelper#format_date` instead of the call to
`Date#strftime`, because we're trying to standardise on using the former. The
behaviour is unchanged.

* I would've liked to have used `SmartAnswer::OutcomeHelper#format_money` for
the weekly SSP amount, but there is a slight difference in behaviour between it
and the call to `Kernel#sprintf` in that the former doesn't display the decimal
places if there are no pence. I plan to ask the content team whether this is
intentional, so that we can standardise this as well.