Fail fast if the ERB template doesn't exist
This will make it harder to forget to add the coverpage or outcome ERB
templates.

If the coverpage or an outcome template is missing then a
`ActionView::MissingTemplate` exception is raised.

Note. Although the `ErbRenderer#content_for` method now returns a single newline
when the `content_for(:symbol)` block doesn't exist, this doesn't affect any
regression test artefacts because of our normalisation of artefacts.