Update content_item_publisher_test
So that it uses the "bridge-of-death" test flow instead of "a-flow-name".

I would've had to add ERB coverpage and outcome templates if I wanted to
continue using "a-flow-name" in this test, and I figured it was better to reuse
an existing test flow rather than create these new files.

I've had to stub `SmartAnswer::FlowRegistry.instance` so that the
`FlowRegistrationPresenter#title` method (which is used in
`FlowContentItem#payload`) returns something sensible. Without the stubbing, we
try to find a real Smart Answer named 'bridge-of-death' which causes the test to
fail.