Updates date_of_year function to subtract any difference in years rather than create new date objects with potentially invalid values.
The date_of_year function allows you to return a date containing the day and month values of the first argument, with the year of the second, which fails when passing a 29th of Feb value for the date and non-leap value for the year.

This change simply subtracts the difference in years from the two arguments, so an earlier year will subtract a positive number of years and a later year subtracts a negative number of years.