Remove unnecessary instantiation of SSP calculator
The `calculator` state variable is already defined (courtesy of the
`next_node_calculation` block below it) when this `calculate` block is invoked.
By making use of that existing state variable, we can DRY up the code somewhat.

I'm particularly interested in doing this, because it reduces the number of
places the `prior_sick_days` state variable is used.