Move prior_sick_days calculation closer to where it's used
I want to make a change to the way the `prev_sick_days` input to the calculator
is calculated, but this change should not affect the logic in the
`start_before_end` validation. This commit effectively duplicates the logic for
calculating `prior_sick_days` in the two places it's used in preparation for
changing the logic in the 2nd occurrence.

Previously the `prior_sick_days` state variable was calculated as part of
processing the response to the `linked_sickness_end_date?` question. This meant
that we could assume that `sick_start_date_for_awe` was set to the response from
the `linked_sickness_start_date?` question and `response` was set to the
response from the `linked_sickness_end_date?` question.

Now in the new occurrence of the logic in the `usual_work_days?` question, we
cannot make the same assumption. Instead we save the response to the
`has_linked_sickness?` question and only assign a meaningful value to
`prior_sick_days` if this response was 'yes'.

Note that all the code that uses the value of this state variable calls `#to_i`
on it, so that if it's not set or set to `nil` (i.e. in the case where there's
no linked period of sickness), then the value is converted to `0`. This is less
than ideal and I plan to improve on it in later commits.