Use use_additional_translation_file in graph_presenter_test
This was the only test that was using `#use_only_translation_file!`. This method
set the `I18n.load_path` to an array containing a single file rather than
appending to it as in `#use_additional_translation_file`. Everywhere else
appends to the `I18n.load_path` so I think that's what we should be doing here
for consistency.