Change regression test responses for SSP
This is a follow-up to the change in questions-and-responses in an earlier
commit which was preparing for the addition of a new validation rule around
the `linked_sickness_end_date?` question.

In that commit one of the responses to the `linked_sickness_end_date?` question
was changed from `2013-09-01` to `2013-03-30`. This change meant that the
`maximum_entitlement_reached` outcome is no longer reached and the relevant
regression test assertion fails.

This commit adds a new response for the `linked_sickness_start_date?` question
in order to reach the `maximum_entitlement_reached`. The new much earlier
response of `2012-09-01` when combined with the responses to the
`linked_sickness_end_date?` question means that we now have some combinations
where the linked period of sickness is sufficiently long that the employee
reaches the maximum SSP entitlement. This means that the
`maximum_entitlement_reached` outcome is reached and the relevant regression
test assertion passes.

I plan to update the regression test artefacts in a subsequent commit.