Update regression test outcome artefacts for SSP
This commit updates the outcome artefacts for calculate-statutory-sick-pay in
the light of the change to the `is_your_employee_getting?` question option from:

    "Ordinary Statutory Paternity Pay" -> "Statutory Paternity Pay"

This string appears inside a conditional paragraph in the body of the main
`entitled_to_sick_pay` outcome.

I ran the following script to make these changes:

    $ find test/artefacts/calculate-statutory-sick-pay -name "*.txt" -exec \
        sed -i '.bak' 's/Ordinary Statutory Paternity Pay/Statutory Paternity Pay/g' {} \;
    $ git clean -fd test/artefacts/calculate-statutory-sick-pay/

The question artefacts were updated separately.