TEMP: Add deprecation warning to MultipleChoice#option
If called with a hash.

Passing a hash to `MultipleChoice#option` allows you to specify the next_node
along with the key for the option. We want to reduce the number of ways that
next nodes can be defined so we're going to remove this functionality. Adding
this deprecation warning should help me identify the places that we're using
this functionality.

I'm testing for the presence of `@flow` in order to set the `flow_name` as we
have some tests that instantiate a `MultipleChoice` object without passing in a
flow. Without this condition, some of these tests would start failing as they
weren't able to show the warning.