Rename regression test artefacts for SSP
This is a follow-up to the change in questions-and-responses in the previous
commit in preparation for the new validation rule.

I'm renaming from `2013-09-01` to `2013-03-30` so that when I run the
regression tests for real I only see actual changes that I'm interested in.

This is the script I used to rename these directories:

    artefacts_path = Rails.root.join('test', 'artefacts', 'calculate-statutory-sick-pay')

    Dir[artefacts_path.join('**/*')].sort_by(&:length).reverse.each do |file_or_directory|
      if File.directory?(file_or_directory)
        if File.basename(file_or_directory) == '2013-09-01'
          old_name = file_or_directory
          new_name = file_or_directory.sub(/2013-09-01$/, '2013-03-30')
          cmd = %(mv #{old_name} #{new_name})
          # puts cmd
          `#{cmd}`
        end
      end
    end