Update regression test artefacts for SSP
This is a follow-up to the change in questions-and-responses in an earlier
commit which was preparing for the addition of a new validation rule around
the `linked_sickness_end_date?` question.

In that commit one of the responses to the `linked_sickness_end_date?` question
was changed from `2013-09-01` to `2013-03-30`.

Having *renamed* the relevant regression test artefacts in the previous commit,
this commit updates the artefacts affected by this new response.

You will notice that all the modified artefacts have a path of the form:

    test/artefacts/calculate-statutory-sick-pay/*/*/*/*/*/yes/2013-02-01/2013-03-30/*.txt

i.e. they all have a linked period of sickness (Q6 has response `yes`) and
they all have the new response to Q6.2 (`2013-03-30`).

Note that no question HTML artefacts have changed.

Moving the end of the linked period of sickness 5 months earlier in these cases
has meant that the employee no longer reaches the maximum entitlement and thus
the outcome changes from `maximum_entitlement_reached` to
`entitled_to_sick_pay`.

Unfortunately there are a lot of changes and I didn't managed to come up with
an automated way to check them, but I have manually sampled the diffs and they
all seem to make sense.

Note that after this change, the `maximum_entitlement_reached` outcome is no
longer reached and the relevant regression test assertion fails. I plan to
address this in a later commit.