Add regression test artefacts for SSP
This is a follow-up to the previous commit in which a new response
(`2012-09-01`) was added to the `linked_sickness_start_date?` question in the
regression test questions-and-responses. This response was added to ensure the
`maximum_entitlement_reached` is reached.

This commit adds the new regression test artefacts generated by this new
response. You should be able to see that all the additions have a path of the
form:

    test/artefacts/calculate-statutory-sick-pay/*/*/*/*/*/yes/2012-09-01/*.txt

i.e. it corresponds to the new response of `2012-09-01`.

Unfortunately there are a large number of new artefacts and I haven't come up
with a good way to automatically check their content. However, I have no reason
to think that the content is wrong and having manually checked a sample subset
of the artefacts, I'm happy that they have the content I would expect.

No question HTML artefacts have changed which is what I expect.

Note that at this point all the regression tests are passing, but I plan to
update the regression test checksums in a subsequent commit.