Rename regression test artefacts for SSP
This commit renames a bunch of regression test artefact directories based on
the renaming of the `ordinary_statutory_paternity_pay` option for the
`is_your_employee_getting?` question.

This option has been renamed from:

    ordinary_statutory_paternity_pay -> statutory_paternity_pay

This commit deals with the renaming of the following combination of checkbox
choices:

    ordinary_statutory_paternity_pay,statutory_adoption_pay

Which becomes:

    statutory_adoption_pay,statutory_paternity_pay

Note the change in order, because the options appear alphabetically.

I used the following script to do the renaming:

    artefacts_path = Rails.root.join('test', 'artefacts', 'calculate-statutory-sick-pay')

    Dir[artefacts_path.join('**/*')].sort_by(&:length).reverse.each do |file_or_directory|
      if File.directory?(file_or_directory)
        if File.basename(file_or_directory) == 'ordinary_statutory_paternity_pay,statutory_adoption_pay'
          old_name = file_or_directory
          new_name = file_or_directory.sub(/ordinary_statutory_paternity_pay,statutory_adoption_pay/, 'statutory_adoption_pay,statutory_paternity_pay')
          cmd = %(mv #{old_name} #{new_name})
          # puts cmd
          `#{cmd}`
        end
      end
    end

I then fixed the one question artefact (which annoyingly has the options in
a different order) using the following command:

    $ git mv test/artefacts/calculate-statutory-sick-pay/ordinary_statutory_paternity_pay,statutory_adoption_pay.html test/artefacts/calculate-statutory-sick-pay/statutory_paternity_pay,statutory_adoption_pay.html