Rename regression test artefacts for SSP
This commit renames a bunch of regression test artefact directories based on
the renaming of the `ordinary_statutory_paternity_pay` option for the
`is_your_employee_getting?` question.

This option has been renamed from:

    ordinary_statutory_paternity_pay -> statutory_paternity_pay

This commit deals with the renaming of the following combination of checkbox
choices:

    additional_statutory_paternity_pay,ordinary_statutory_paternity_pay

Which becomes:

    additional_statutory_paternity_pay,statutory_paternity_pay

There is no change in order in this case.

I used the following script to do the renaming:

    artefacts_path = Rails.root.join('test', 'artefacts', 'calculate-statutory-sick-pay')

    Dir[artefacts_path.join('**/*')].sort_by(&:length).reverse.each do |file_or_directory|
      if File.directory?(file_or_directory)
        if File.basename(file_or_directory) == 'additional_statutory_paternity_pay,ordinary_statutory_paternity_pay'
          old_name = file_or_directory
          new_name = file_or_directory.sub(/additional_statutory_paternity_pay,ordinary_statutory_paternity_pay/, 'additional_statutory_paternity_pay,statutory_paternity_pay')
          cmd = %(mv #{old_name} #{new_name})
          # puts cmd
          `#{cmd}`
        end
      end
    end